🚚 Footer style 코드 별도의 파일로 관리
