+++ coz-shopping/src/App.tsx
@@ -1,6 +1,6 @@
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import Header from "./components/Header/Header";
-import Footer from "./components/Footer";
+import Footer from "./components/Footer/Footer";
 import Main from "./components/Home/Main";
 import ProductList from "./components/ProductList/ProductList";
 import BookMarkList from "./components/BookMarkList/BookMarkList";
