+++ coz-shopping/src/components/Footer/Footer.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import { FooterWrapper, FooterContent } from "./FooterStyle";
+
+function Footer() {
+  return (
+    <FooterWrapper>
+      <FooterContent>개인정보 처리방침 | 이용약관</FooterContent>
+      <FooterContent>All rights reserved @ CodeStates</FooterContent>
+    </FooterWrapper>
+  );
+}
+
+export default Footer;
